experimenting with the update function
it turns out that the `dt` parameter is not a constant number

and the variation of it "seems" random.

and also this code sometimes causes the shapes to just disappear. Also setting the gravity to a very big number results in invisible / missing shapes too.

I am beginning to wonder if python is not the right tool for the job. Even though it's the perfect tool for data science... not necessarily for fun physics-based simulations.

I might end up turning off the fancy gravity + friction + elasticity altogether and just allow for simple collisions.